Fix proto schema client generation
