+++ tests/proto_build_test/src/client.rs
@@ -17,43 +17,20 @@ pub mod extra_types {
 
     #[proto_message]
     pub struct BuildRequest {
-        #[proto(tag = 1)]
         pub config: BuildConfig,
-        #[proto(tag = 2)]
         pub ping: RizzPing,
-        #[proto(tag = 3)]
         pub owner: Id,
     }
 
     #[proto_message]
     pub struct BuildResponse {
-        #[proto(tag = 1)]
         pub status: ServiceStatus,
-        #[proto(tag = 2)]
-        pub envelope: Envelope,
+        pub envelope: Envelope<GoonPong>,
     }
 
     #[proto_message]
     pub struct Envelope<T> {
-        #[proto(tag = 1)]
-        pub payload: BuildRequest,
-        #[proto(tag = 2)]
-        pub trace_id: ::proto_rs::alloc::string::String,
-    }
-
-    #[proto_message]
-    pub struct Envelope<T> {
-        #[proto(tag = 1)]
-        pub payload: BuildResponse,
-        #[proto(tag = 2)]
-        pub trace_id: ::proto_rs::alloc::string::String,
-    }
-
-    #[proto_message]
-    pub struct Envelope<T> {
-        #[proto(tag = 1)]
-        pub payload: GoonPong,
-        #[proto(tag = 2)]
+        pub payload: T,
         pub trace_id: ::proto_rs::alloc::string::String,
     }
 
@@ -63,37 +40,11 @@ pub mod fastnum {
     use proto_rs::{proto_message, proto_rpc};
 
     #[proto_message]
-    pub struct D128Proto {
-        #[proto(tag = 1)]
+    pub struct D128 {
         pub lo: u64,
-        #[proto(tag = 2)]
         pub hi: u64,
-        #[proto(tag = 3)]
         pub fractional_digits_count: i32,
-        #[proto(tag = 4)]
         pub is_negative: bool,
-    }
-
-    #[proto_message]
-    pub struct D128Proto {
-        #[proto(tag = 1)]
-        pub lo: u64,
-        #[proto(tag = 2)]
-        pub hi: u64,
-        #[proto(tag = 3)]
-        pub fractional_digits_count: i32,
-        #[proto(tag = 4)]
-        pub is_negative: bool,
-    }
-
-    #[proto_message]
-    pub struct UD128Proto {
-        #[proto(tag = 1)]
-        pub lo: u64,
-        #[proto(tag = 2)]
-        pub hi: u64,
-        #[proto(tag = 3)]
-        pub fractional_digits_count: i32,
     }
 
 }
@@ -111,23 +62,18 @@ pub mod goon_types {
 
     #[proto_message]
     pub struct GoonPong {
-        #[proto(tag = 1)]
         pub id: Id,
-        #[proto(tag = 2)]
         pub status: ServiceStatus,
     }
 
     #[proto_message]
     pub struct Id {
-        #[proto(tag = 1)]
         pub id: u64,
     }
 
     #[proto_message]
     pub struct RizzPing {
-        #[proto(tag = 1)]
         pub id: Id,
-        #[proto(tag = 2)]
         pub status: ServiceStatus,
     }
 
@@ -146,54 +92,45 @@ pub mod rizz_types {
 pub mod sigma_rpc_simple {
     #[allow(unused_imports)]
     use proto_rs::{proto_message, proto_rpc};
+    use crate::extra_types::BuildRequest;
     use crate::extra_types::BuildResponse;
     use crate::extra_types::Envelope;
-    use crate::fastnum::D128Proto;
-    use crate::fastnum::UD128Proto;
+    use crate::fastnum::D128;
     use crate::goon_types::GoonPong;
     use crate::goon_types::Id;
     use crate::goon_types::RizzPing;
     use crate::rizz_types::BarSub;
     use crate::rizz_types::FooResponse;
+    use fastnum::UD128;
 
     #[proto_rpc(rpc_package = "sigma_rpc", rpc_server = false, rpc_client = true)]
     pub trait SigmaRpc {
-        type RizzUniStream: ::tonic::codegen::tokio_stream::Stream<Item = ::core::result::Result<FooResponse, ::tonic::Status>> + ::core::marker::Send + 'static;
+        type RizzUniStream: ::tonic::codegen::tokio_stream::Stream<Item = ::core::result::Result<FooResponse, ::tonic::Status>> + ::core::marker::Send;
 
         async fn rizz_ping(
             &self,
             request: ::tonic::Request<RizzPing>,
-        ) -> ::core::result::Result<::tonic::Response<GoonPong>, ::tonic::Status>
-        where
-            Self: ::core::marker::Send + ::core::marker::Sync;
+        ) -> ::core::result::Result<::tonic::Response<GoonPong>, ::tonic::Status>;
 
         async fn rizz_uni(
             &self,
             request: ::tonic::Request<BarSub>,
-        ) -> ::core::result::Result<::tonic::Response<Self::RizzUniStream>, ::tonic::Status>
-        where
-            Self: ::core::marker::Send + ::core::marker::Sync;
+        ) -> ::core::result::Result<::tonic::Response<Self::RizzUniStream>, ::tonic::Status>;
 
         async fn build(
             &self,
-            request: ::tonic::Request<Envelope>,
-        ) -> ::core::result::Result<::tonic::Response<Envelope>, ::tonic::Status>
-        where
-            Self: ::core::marker::Send + ::core::marker::Sync;
+            request: ::tonic::Request<Envelope<BuildRequest>>,
+        ) -> ::core::result::Result<::tonic::Response<Envelope<BuildResponse>>, ::tonic::Status>;
 
         async fn owner_lookup(
             &self,
             request: ::tonic::Request<Id>,
-        ) -> ::core::result::Result<::tonic::Response<BuildResponse>, ::tonic::Status>
-        where
-            Self: ::core::marker::Send + ::core::marker::Sync;
+        ) -> ::core::result::Result<::tonic::Response<BuildResponse>, ::tonic::Status>;
 
         async fn test_decimals(
             &self,
-            request: ::tonic::Request<UD128Proto>,
-        ) -> ::core::result::Result<::tonic::Response<D128Proto>, ::tonic::Status>
-        where
-            Self: ::core::marker::Send + ::core::marker::Sync;
+            request: ::tonic::Request<UD128>,
+        ) -> ::core::result::Result<::tonic::Response<D128>, ::tonic::Status>;
 
     }
 
@@ -203,25 +140,22 @@ pub mod solana {
     use proto_rs::{proto_message, proto_rpc};
 
     #[proto_message]
-    pub struct AddressProto {
-        #[proto(tag = 1)]
-        pub inner: ::proto_rs::alloc::vec::Vec<u8>,
+    pub struct Address {
+        pub inner: [u8; BYTES],
     }
 
     #[proto_message]
-    pub struct KeypairProto {
-        #[proto(tag = 1)]
-        pub inner: ::proto_rs::alloc::vec::Vec<u8>,
+    pub struct Keypair {
+        pub inner: [u8; BYTES],
     }
 
     #[proto_message]
-    pub struct SignatureProto {
-        #[proto(tag = 1)]
-        pub inner: ::proto_rs::alloc::vec::Vec<u8>,
+    pub struct Signature {
+        pub inner: [u8; BYTES],
     }
 
     #[proto_message]
-    pub enum InstructionErrorProto {
+    pub enum InstructionError {
         GenericError,
         InvalidArgument,
         InvalidInstructionData,
@@ -248,7 +182,6 @@ pub mod solana {
         AccountBorrowOutstanding,
         DuplicateAccountOutOfSync,
         Custom(
-            #[proto(tag = 1)]
             u32,
         ),
         InvalidError,
@@ -282,7 +215,7 @@ pub mod solana {
     }
 
     #[proto_message]
-    pub enum TransactionErrorProto {
+    pub enum TransactionError {
         AccountInUse,
         AccountLoadedTwice,
         AccountNotFound,
@@ -292,10 +225,8 @@ pub mod solana {
         AlreadyProcessed,
         BlockhashNotFound,
         InstructionError {
-            #[proto(tag = 1)]
             index: u32,
-            #[proto(tag = 2)]
-            error: InstructionErrorProto,
+            error: InstructionError,
         },
         CallChainTooDeep,
         MissingSignatureForFee,
@@ -319,18 +250,15 @@ pub mod solana {
         WouldExceedMaxVoteCostLimit,
         WouldExceedAccountDataTotalLimit,
         DuplicateInstruction(
-            #[proto(tag = 1)]
             u32,
         ),
         InsufficientFundsForRent {
-            #[proto(tag = 1)]
             account_index: u32,
         },
         MaxLoadedAccountsDataSizeExceeded,
         InvalidLoadedAccountsDataSizeLimit,
         ResanitizationNeeded,
         ProgramExecutionTemporarilyRestricted {
-            #[proto(tag = 1)]
             account_index: u32,
         },
         UnbalancedTransaction,
